Crate substrate_application_crypto

Traits and macros for constructing application specific strongly typed crypto wrappers.

Modules

ed25519
Ed25519 crypto types.

key_types
Known key types; this also functions as a global registry of key types for projects wishing to avoid collisions with each other.

sr25519
Sr25519 crypto types.

Macros

app_crypto
Declares Public, Pair, Signature types which are functionally equivalent to $pair, but are new Application-specific types whose identifier is $key_type.

wrap
Implement bidirectional From and on-way AsRef/AsMut for two types, $inner and $outer.

Structs

KeyTypeId
An identifier for a type of cryptographic key.

Traits

AppKey
An application-specific key.

AppPair
A application's key pair.

AppPublic
A application's public key.

AppSignature
A application's signature.

MaybeHash
Type which implements Hash in std, not when no-std (std variant).

RuntimeAppPublic
A runtime interface for an application's public key.

RuntimePublic
A runtime interface for a public key.