Struct substrate_application_crypto::sr25519::Signature

pub struct Signature(pub [u8; 64]);

An Schnorrkel/Ristretto x25519 ("sr25519") signature.

Instead of importing it for the local module, alias it to be available as a public type

Methods

impl Signature

pub fn from_raw(data: [u8; 64]) -> Signature

A new instance from the given 64-byte data.

NOTE: No checking goes on to ensure this is a real signature. Only use it if you are certain that the array actually is a signature, or if you immediately verify the signature. All functions that verify signatures will fail if the Signature is not actually a valid signature.

pub fn from_slice(data: &[u8]) -> Signature

A new instance from the given slice that should be 64 bytes long.

NOTE: No checking goes on to ensure this is a real signature. Only use it if you are certain that the array actually is a signature. GIGO!

pub fn from_h512(v: H512) -> Signature

A new instance from an H512.

NOTE: No checking goes on to ensure this is a real signature. Only use it if you are certain that the array actually is a signature. GIGO!